Changed from alpine distro to perl-slim
Tried to compile perl on Alpine Linux, but that's too hard due C
libraries available.
But decided that running on a perl with ithreads enabled was a waste of
CPU and memory, so moved to perl-slim image with is compiled without
ithreads enabled.
Using multi-stage build to avoid using more space than necessary.
Updated the documentation as well.